+++ CHANGES.rst
@@ -3,7 +3,10 @@ Changelog of lizard-nxt client
 
 Unreleased (2.9.0) (XXXX-XX-XX)
 -------------------------------
--
+
+- Remove layergroups from search.
+
+- Adapt search results to updated full-text search API response.
 
 
 Release 2.9.4 (2016-3-1)